NetworkDistances

Documentation for NetworkDistances.

NetworkDistances.Normalised

NetworkDistances.Precomputed

NetworkDistances.diffusion_dist

NetworkDistances.emd

NetworkDistances.emd2

NetworkDistances.get_cost_matrix_dynamic

NetworkDistances.get_cost_matrix_dynamic

NetworkDistances.get_cost_matrix_fixed

NetworkDistances.get_cost_matrix_fixed

NetworkDistances.get_info_deep

NetworkDistances.hamming_dist

NetworkDistances.jaccard_dist

NetworkDistances.pre_compute

NetworkDistances.progress_pairwise

StatsAPI.pairwise

StatsAPI.pairwise!

StatsAPI.pairwise!

StatsAPI.pairwise!

NetworkDistances.Normalised — Type

Normalised distance obtained by applying Steinhaus transform to given distance.

Note this will be a metric given d is a metric.

NetworkDistances.Precomputed — Type

Distance which just looks-up a cache of stored distances.

NetworkDistances.diffusion_dist — Method

Compute the diffusion distance between two adjacency matrices with t the time of diffusion.

NetworkDistances.emd — Method

Wrapper for ot.emd() method of POT python package.

NetworkDistances.emd2 — Method

Wrapper for ot.emd2() method of POT python package.

NetworkDistances.get_cost_matrix_dynamic — Method

get_cost_matrix_dynamic(d::AbstractMatchingDistance, X, Y)

Return cost matrix for evaluating matching-based distance between X and Y.

This is done "dynamically" in the sense that the first axis of the output matrix will always correspond to the shorter of the two passed vectors X and Y.

NetworkDistances.get_cost_matrix_fixed — Method

get_cost_matrix_fixed(d::AbstractMatchingDistance, X, Y)

Return cost matrix for evaluating matching-based distance between X and Y.

This is "fixed" in the sense the first axis of output matrix always corresponds to X and the second to Y.

NetworkDistances.get_info_deep — Method

get_info_deep(d::MatchingBasedDistance, X::Vector{Vector{S}}, Y::Vector{Vector{S}})

Returns two vectors indicating which entries have been matched. An entry being true in the output implies this entry was matched with one in the other observation.

NetworkDistances.hamming_dist — Method

Compute the Hamming distance between two adjacency matrices (directed graphs)

NetworkDistances.jaccard_dist — Method

Compute the Jaccard distance between two adjacency matrices (directed graphs)

NetworkDistances.pre_compute — Method

Constructor method which takes distance + collection of values and makes a pre-computed distance.

NetworkDistances.progress_pairwise — Method

Distance matrix calculation between elements of Vectors with progress bar.

StatsAPI.pairwise! — Method

In-place distance matrix calculation between elements of Vectors.

StatsAPI.pairwise! — Method

In-place distance matrix calculation between elements of Vectors. (For SubArray)

StatsAPI.pairwise! — Method

In-place distance matrix calculation between elements of Vectors. (For SubArray)

StatsAPI.pairwise — Method

Distance matrix calculation between elements of Vectors. This is a custom extension of the function in the Distances.jl package to allow vectors of general type. The function in Distances.jl is designed for univariate/multivariate data and so takes as input either vectors or matrices (data points as rows).