Matching distances

Type hierarchy

At the top level we have

AbstractMatchingDistance <: SemiMetric

This encompasses all matching-based distances. Within this, we have two further abstract classes

CompleteMatchingDistance <: AbstractMatchingDistance - these are the complete matching distances which optimise over complete matchings only (i.e. where all path from smaller observation are included in matching);

General{T} <: AbstractMatchingDistance where T <: CompleteMatchingDistance - these can have paths from both observations being unmatched. These are defined given complete matchings via General(d) where d::CompleteMatchingDistance.

Other methods

NetworkDistances.get_cost_matrix_dynamic — Function

get_cost_matrix_dynamic(d::AbstractMatchingDistance, X, Y)

Return cost matrix for evaluating matching-based distance between X and Y.

This is done "dynamically" in the sense that the first axis of the output matrix will always correspond to the shorter of the two passed vectors X and Y.

NetworkDistances.get_cost_matrix_fixed — Function

get_cost_matrix_fixed(d::AbstractMatchingDistance, X, Y)

Return cost matrix for evaluating matching-based distance between X and Y.

This is "fixed" in the sense the first axis of output matrix always corresponds to X and the second to Y.